Split config.ts
