+++ packages/core/src/core/config/config-type.error.ts
@@ -1,0 +1,9 @@
+export class ConfigTypeError extends Error {
+  constructor(readonly key: string, readonly expected: string, readonly actual: string) {
+    super(
+      `The value of the configuration key "${key}" has an invalid type.\n`
+      + '\n'
+      + `Expected a "${expected}", but got a "${actual}".`
+    );
+  }
+}
