+++ packages/core/src/core/config/index.ts
@@ -1,2 +1,4 @@
-export { Config, ConfigNotFoundError, ConfigTypeError } from './config';
 export { ConfigMock } from './config-mock';
+export { ConfigNotFoundError } from './config-not-found.error';
+export { ConfigTypeError } from './config-type.error';
+export { Config } from './config';
