+++ packages/core/src/core/config/utils.ts
@@ -1,0 +1,6 @@
+export function dotToUnderscore(str: string): string {
+  return str
+    .replace(/([A-Z])/g, letter => `_${letter}`)
+    .replace(/\./g, '_')
+    .toUpperCase();
+}
