+++ packages/core/src/core/index.ts
@@ -3,5 +3,5 @@ export { createController } from './controllers';
 export * from './http';
 export * from './hooks';
 export * from './routes';
-export { Config, ConfigMock, ConfigNotFoundError, ConfigTypeError } from './config';
+export * from './config';
 export { createService, dependency, ServiceManager } from './service-manager';
